pyproject.toml
[project]
name = "sonarqube-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=3.1.1",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "respx>=0.22",
]

[project.scripts]
sonarqube-mcp = "sonarqube_mcp.main:sonarqube_mcp"
